move spinlock into mod.rs
